@@ -102,31 +105,39 @@ include_directories(${fts_server_SOURCE_DIR}/../..
                    )
 
 # Add necessary defines:
-add_definitions(-DD_STRING_MYSQL
-                -DD_DSTRING
+add_definitions(-DD_DSTRING
                 -DD_COMPILES_SERVER
                 -DD_NOCEGUI
                 -DD_OWN_ENDIAN
                 -DSDL_BYTEORDER=SDL_LIL_ENDIAN)
 
+if(MSVC)
+    add_definitions(-D_CRT_SECURE_NO_WARNINGS
+                    -D_WINSOCK_DEPRECATED_NO_WARNINGS
+    )
+endif(MSVC)
+
 # Compiler-dependent and build-dependend flags:
 if(CMAKE_COMPILER_IS_GNUCC OR CMAKE_COMPILER_IS_GNUCXX)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x -Wall")
-    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -std=c++0x -Wall -pg")
-    set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} -std=c++0x -Wall")
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++14 -Wall -fpermissive")
+    set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -std=c++14 -Wall -pg")
+    set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} -std=c++14 -Wall")
+    # And show the flags.
+    if(CMAKE_BUILD_TYPE STREQUAL "Debug")
+        message(STATUS "CXX Flags (Debug) are: ${CMAKE_CXX_FLAGS_DEBUG}")
+    else()
+        message(STATUS "CXX Flags (Release) are: ${CMAKE_CXX_FLAGS_RELEASE}")
+    endif()
 elseif(MSVC)
+    link_directories( ${MYSQL_LIBRARIES_DIR}  )	
+    set( MYSQL_LIBRARIES "libmysql.lib;mysqlclient.lib" )
 endif()
 
-# And show the flags.
-if(CMAKE_BUILD_TYPE STREQUAL "Debug")
-    message(STATUS "CXX Flags (Debug) are: ${CMAKE_CXX_FLAGS_DEBUG}")
-else()
-    message(STATUS "CXX Flags (Release) are: ${CMAKE_CXX_FLAGS_RELEASE}")
-endif()
 
 # The compiling process. #
 ##########################
+
 add_executable(fts-server ${SOURCES})
-target_link_libraries(fts-server ${MYSQL_LIBRARIES} ${CMAKE_THREAD_LIBS_INIT})
+target_link_libraries(fts-server ${MYSQL_LIBRARIES} ${CMAKE_THREAD_LIBS_INIT} )
 
 install(TARGETS fts-server DESTINATION bin)
