@@ -1,3 +1,6 @@
+
+#include <algorithm>
+
 #include "channel.h"
 #include "client.h"
 #include "db.h"
@@ -8,11 +11,7 @@
 using namespace FTS;
 using namespace FTSSrv2;
 
-FTSSrv2::Channel::Channel( int in_iID,
-                    bool in_bPublic,
-                    const String & in_sName,
-                    const String & in_sMotto,
-                    const String & in_sAdmin )
+FTSSrv2::Channel::Channel( int in_iID, bool in_bPublic, const String & in_sName, const String & in_sMotto, const String & in_sAdmin )
     : m_iID(in_iID)
     , m_bPublic(in_bPublic)
     , m_sName(in_sName)
@@ -29,7 +28,7 @@ FTSSrv2::Channel::~Channel( )
 {
     // Kick all users that are still in the channel.
     while(!m_lpUsers.empty()) {
-        this->kick(NULL, m_lpUsers.front()->getNick());
+        this->kick(nullptr, m_lpUsers.front()->getNick());
     }
 
     m_lsOperators.clear();
@@ -45,18 +44,18 @@ int FTSSrv2::Channel::join(Client *in_pUser)
         m_lpUsers.push_back(in_pUser);
 
         // Tell everybody that somebody joined.
-        for(std::list<Client *>::const_iterator i = m_lpUsers.begin() ; i != m_lpUsers.end() ; ++i) {
+        for(const auto& i : m_lpUsers) {
             // But don't tell myself !
-            if( in_pUser == (*i) )
+            if( in_pUser == i )
                 continue;
 
-            (*i)->sendChatJoins( sJoinName );
+            i->sendChatJoins( sJoinName );
         }
     }
 
     in_pUser->setMyChannel(this);
 
-    FTSMSGDBG(in_pUser->getNick() + " joined " + m_sName, 1);
+    FTSMSGDBG(in_pUser->getNick() + " joined " + m_sName, 4);
     return ERR_OK;
 }
 
@@ -70,14 +69,14 @@ int FTSSrv2::Channel::quit(Client *in_pUser)
         m_lpUsers.remove(in_pUser);
 
         // Tell everybody that somebody left.
-        for(std::list<Client *>::const_iterator i = m_lpUsers.begin() ; i != m_lpUsers.end() ; ++i) {
-            (*i)->sendChatQuits(sQuitName);
+        for(const auto& i : m_lpUsers) {
+            i->sendChatQuits(sQuitName);
         }
     }
 
-    in_pUser->setMyChannel(NULL);
+    in_pUser->setMyChannel(nullptr);
 
-    FTSMSGDBG(in_pUser->getNick() + " left " + m_sName, 1);
+    FTSMSGDBG(in_pUser->getNick() + " left " + m_sName, 4);
     return ERR_OK;
 }
 
@@ -101,7 +100,7 @@ int FTSSrv2::Channel::save()
 
     // Or just save modifications ?
     } else {
-        sQuery = "UPDATE `"DSRV_TBL_CHANS"`"
+        sQuery = "UPDATE `" DSRV_TBL_CHANS "`"
                  " SET `"+DataBase::getUniqueDB()->TblChansField(DSRV_TBL_CHANS_NAME)+"`"
                         "=\'" + DataBase::getUniqueDB()->escape(m_sName) + "\',"
                       "`"+DataBase::getUniqueDB()->TblChansField(DSRV_TBL_CHANS_MOTTO)+"`"
@@ -120,8 +119,8 @@ int FTSSrv2::Channel::save()
     }
 
     // Now we have to update the operators table.
-    for(std::list<String>::const_iterator i = m_lsOperators.begin() ; i != m_lsOperators.end() ; ++i) {
-        sQuery = "\'" + DataBase::getUniqueDB()->escape((*i)) + "\'," + String::nr(m_iID);
+    for(const auto& i : m_lsOperators) {
+        sQuery = "\'" + DataBase::getUniqueDB()->escape(i) + "\'," + String::nr(m_iID);
         DataBase::getUniqueDB()->storedFunctionInt( "channelAddOp", sQuery );
     }
 
@@ -145,20 +144,20 @@ int FTSSrv2::Channel::op(const String & in_sUser, bool in_bOninit)
     // Don't check all these things during the startup.
     if(!in_bOninit) {
         // Check if the user is in the channel.
-        if(NULL == this->getUserIfPresent(in_sUser)) {
-            FTSMSGDBG(in_sUser+" does not exist in this channel", 1);
+        if(nullptr == this->getUserIfPresent(in_sUser)) {
+            FTSMSGDBG(in_sUser+" does not exist in this channel", 4);
             return -31;
         }
 
         // Do not add the user to the operators list if he already is.
         if(this->isop(in_sUser)) {
-            FTSMSGDBG(in_sUser+" is already operator in this channel", 1);
+            FTSMSGDBG(in_sUser+" is already operator in this channel", 4);
             return -32;
         }
     }
 
     if(in_sUser == m_sAdmin) {
-        FTSMSGDBG(in_sUser+" is the channel admin!", 1);
+        FTSMSGDBG(in_sUser+" is the channel admin!", 4);
         return -33;
     }
 
@@ -169,8 +168,8 @@ int FTSSrv2::Channel::op(const String & in_sUser, bool in_bOninit)
 
         if(!in_bOninit) {
             // Tell everybody that somebody op's, including to me.
-            for(std::list<Client *>::const_iterator i = m_lpUsers.begin() ; i != m_lpUsers.end() ; ++i) {
-                (*i)->sendChatOped( in_sUser );
+            for(const auto& i : m_lpUsers) {
+                i->sendChatOped( in_sUser );
             }
         }
     }
@@ -178,7 +177,7 @@ int FTSSrv2::Channel::op(const String & in_sUser, bool in_bOninit)
     // Store the changes in the database.
     this->save();
 
-    FTSMSGDBG(in_sUser+" op in "+m_sName, 1);
+    FTSMSGDBG(in_sUser+" op in "+m_sName, 4);
     return ERR_OK;
 }
 
@@ -186,7 +185,7 @@ int FTSSrv2::Channel::deop( const String & in_sUser )
 {
     // Check if the user is op.
     if(!this->isop(in_sUser)) {
-        FTSMSGDBG(in_sUser+" is not operator in this channel", 1);
+        FTSMSGDBG(in_sUser+" is not operator in this channel", 4);
         return -1;
     }
 
@@ -196,23 +195,23 @@ int FTSSrv2::Channel::deop( const String & in_sUser )
         m_lsOperators.remove(in_sUser);
 
         // Tell everybody that somebody Deop's, including to me.
-        for(std::list<Client *>::const_iterator i = m_lpUsers.begin() ; i != m_lpUsers.end() ; ++i) {
-            (*i)->sendChatDeOped( in_sUser );
+        for(const auto& i : m_lpUsers) {
+            i->sendChatDeOped( in_sUser );
         }
     }
 
     // Store the changes in the database.
     this->save();
 
-    FTSMSGDBG(in_sUser+" deop in "+m_sName, 1);
+    FTSMSGDBG(in_sUser+" deop in "+m_sName, 4);
     return ERR_OK;
 }
 
 bool FTSSrv2::Channel::isop( const String & in_sUser )
 {
     Lock l(m_mutex);
-    for(std::list<String>::const_iterator i = m_lsOperators.begin() ; i != m_lsOperators.end() ; ++i) {
-        if(in_sUser.ieq(*i)) {
+    for(const auto& i : m_lsOperators) {
+        if(in_sUser.ieq(i)) {
             return true;
         }
     }
@@ -228,7 +227,7 @@ int FTSSrv2::Channel::messageToAll( const Client &in_From, const String & in_sMe
     p->append(in_From.getNick());
     p->append(in_sMessage);
 
-    FTSMSGDBG(in_From.getNick()+" says "+in_sMessage, 1);
+    FTSMSGDBG(in_From.getNick()+" says "+in_sMessage, 4);
 
     this->sendPacketToAll(p);
     SAFE_DELETE(p);
@@ -248,8 +247,8 @@ Packet *FTSSrv2::Channel::makeSystemMessagePacket( const String &in_sMessageID )
 int FTSSrv2::Channel::sendPacketToAll( Packet *in_pPacket )
 {
     Lock l(m_mutex);
-    for(std::list<Client *>::const_iterator i = m_lpUsers.begin() ; i != m_lpUsers.end() ; ++i) {
-        (*i)->sendPacket(in_pPacket);
+    for(const auto& i : m_lpUsers) {
+        i->sendPacket(in_pPacket);
     }
 
     return ERR_OK;
@@ -266,8 +265,8 @@ int FTSSrv2::Channel::setMotto( const String & in_sMotto, const String & in_sUse
         m_sMotto = in_sMotto;
 
         // Tell everybody that somebody changed the motto, including to me.
-        for(std::list<Client *>::const_iterator i = m_lpUsers.begin() ; i != m_lpUsers.end() ; ++i) {
-            (*i)->sendChatMottoChanged( in_sUser, in_sMotto );
+        for(const auto& i : m_lpUsers) {
+            i->sendChatMottoChanged( in_sUser, in_sMotto );
         }
     }
 
@@ -277,10 +276,10 @@ int FTSSrv2::Channel::setMotto( const String & in_sMotto, const String & in_sUse
 int FTSSrv2::Channel::kick( const Client *in_pFrom, const String & in_sUser )
 {
     String sKicker;
-    Client *pKicked = NULL;
+    Client *pKicked = nullptr;
 
     // If the system wants to kick him out.
-    if(in_pFrom == NULL) {
+    if(in_pFrom == nullptr) {
         sKicker = "Arkana-FTS Server";
     } else {
         sKicker = in_pFrom->getNick();
@@ -304,7 +303,7 @@ int FTSSrv2::Channel::kick( const Client *in_pFrom, const String & in_sUser )
     pKicked = this->getUserIfPresent(in_sUser);
 
     // User not in the channel ?
-    if(pKicked == NULL)
+    if(pKicked == nullptr)
         return -21;
 
     // Kick him out to the default channel.
@@ -330,291 +329,11 @@ int FTSSrv2::Channel::kick( const Client *in_pFrom, const String & in_sUser )
 Client *FTSSrv2::Channel::getUserIfPresent(const String &in_sUsername)
 {
     Lock l(m_mutex);
-    Client *pCli = NULL;
-    for(std::list<Client *>::const_iterator i = m_lpUsers.begin() ; i != m_lpUsers.end() ; ++i) {
-        pCli = *i;
+    for(const auto& pCli : m_lpUsers) {
         if(pCli->getNick().ieq(in_sUsername)) {
             return pCli;
         }
     }
 
-    return NULL;
-}
-
-FTSSrv2::ChannelManager::ChannelManager()
-{
-    m_lpChannels.clear();
-}
-
-FTSSrv2::ChannelManager::~ChannelManager()
-{
-    this->saveChannels( );
-
-    for( std::list<FTSSrv2::Channel *>::const_iterator i = m_lpChannels.begin() ;
-         i != m_lpChannels.end() ; i++ ) {
-        delete (*i);
-    }
-
-    m_lpChannels.clear();
-}
-
-static FTSSrv2::ChannelManager *g_pTheCM = NULL;
-
-int FTSSrv2::ChannelManager::init()
-{
-    g_pTheCM = new FTSSrv2::ChannelManager();
-
-    // First, load all the channels from the database.
-    if(ERR_OK != g_pTheCM->loadChannels()) {
-        SAFE_DELETE(g_pTheCM);
-        return -1;
-    }
-
-    // Then, check if the main channel is existing.
-    if(g_pTheCM->getDefaultChannel() == NULL) {
-        // If not, create it with Pompei2 as admin.
-        // We do create it manually here because it's a special case.
-        g_pTheCM->m_lpChannels.push_back(new FTSSrv2::Channel(-1, true,
-                                                      DSRV_DEFAULT_CHANNEL_NAME,
-                                                      DSRV_DEFAULT_CHANNEL_MOTTO,
-                                                      DSRV_DEFAULT_CHANNEL_ADMIN));
-
-    } else if(g_pTheCM->getDefaultChannel()->getAdmin() != DSRV_DEFAULT_CHANNEL_ADMIN) {
-        // Or if somehow another admin is entered in it, set it to the default admin!
-        g_pTheCM->getDefaultChannel()->setAdmin(DSRV_DEFAULT_CHANNEL_ADMIN);
-    }
-
-    // Same for the dev's channel.
-    if(g_pTheCM->findChannel(DSRV_DEVS_CHANNEL_NAME) == NULL) {
-        // If not, create it with Pompei2 as admin.
-        // We do create it manually here because it's a special case.
-        g_pTheCM->m_lpChannels.push_back(new FTSSrv2::Channel(-1, true,
-                                                      DSRV_DEVS_CHANNEL_NAME,
-                                                      DSRV_DEVS_CHANNEL_MOTTO,
-                                                      DSRV_DEVS_CHANNEL_ADMIN));
-
-    } else if(g_pTheCM->findChannel(DSRV_DEVS_CHANNEL_NAME)->getAdmin() != DSRV_DEVS_CHANNEL_ADMIN) {
-        // Or if somehow another admin is entered in it, set it to the default admin!
-        g_pTheCM->findChannel(DSRV_DEVS_CHANNEL_NAME)->setAdmin(DSRV_DEVS_CHANNEL_ADMIN);
-    }
-
-    return ERR_OK;
-}
-
-FTSSrv2::ChannelManager *FTSSrv2::ChannelManager::getManager()
-{
-    return g_pTheCM;
-}
-
-int FTSSrv2::ChannelManager::deinit()
-{
-    SAFE_DELETE(g_pTheCM);
-    return ERR_OK;
-}
-
-int FTSSrv2::ChannelManager::loadChannels(void)
-{
-    MYSQL_RES *pRes = NULL;
-    MYSQL_ROW pRow = NULL;
-
-    // Do the query to get the field.
-    String sQuery = "SELECT  `"+DataBase::getUniqueDB()->TblChansField(DSRV_TBL_CHANS_ID)+"`"
-                            ",`"+DataBase::getUniqueDB()->TblChansField(DSRV_TBL_CHANS_PUBLIC)+"`"
-                            ",`"+DataBase::getUniqueDB()->TblChansField(DSRV_TBL_CHANS_NAME)+"`"
-                            ",`"+DataBase::getUniqueDB()->TblChansField(DSRV_TBL_CHANS_MOTTO)+"`"
-                            ",`"+DataBase::getUniqueDB()->TblChansField(DSRV_TBL_CHANS_ADMIN)+"`"
-                     " FROM `"DSRV_TBL_CHANS"`";
-
-    if(!DataBase::getUniqueDB()->query(pRes, sQuery)) {
-        DataBase::getUniqueDB()->free(pRes);
-        return -1;
-    }
-
-    // Invalid record ? forget about it!
-    if(pRes == NULL || mysql_num_fields(pRes) < 5) {
-        DataBase::getUniqueDB()->free(pRes);
-        return -2;
-    }
-
-    // Create every single channel.
-    while(NULL != (pRow = mysql_fetch_row(pRes))) {
-        int iChannelID = atoi(pRow[0]);
-        bool bPublic = (pRow[1] == NULL ? false : (pRow[1][0] == '0' ? false : true));
-        String sChanName = pRow[2];
-        String sChanMotto = pRow[3];
-        String sChanAdmin = pRow[4];
-
-        FTSSrv2::Channel *pChan = new FTSSrv2::Channel(iChannelID, bPublic, sChanName, sChanMotto, sChanAdmin);
-        m_lpChannels.push_back(pChan);
-    }
-
-    DataBase::getUniqueDB()->free(pRes);
-
-    // Now read all channel operators.
-    sQuery = "SELECT  `"+DataBase::getUniqueDB()->TblChanOpsField(DSRV_VIEW_CHANOPS_NICK)+"`"
-                    ",`"+DataBase::getUniqueDB()->TblChanOpsField(DSRV_VIEW_CHANOPS_CHAN)+"`"
-             " FROM `"DSRV_VIEW_CHANOPS"`";
-    if(!DataBase::getUniqueDB()->query(pRes, sQuery)) {
-        DataBase::getUniqueDB()->free(pRes);
-        return -3;
-    }
-
-    // Invalid record ? forget about it!
-    if(pRes == NULL || mysql_num_fields(pRes) < 2) {
-        DataBase::getUniqueDB()->free(pRes);
-        return -4;
-    }
-
-    // Setup every operator<->channel connection.
-    // But first just put all assocs. in a list because we need to free the DB.
-    std::list<std::pair<FTSSrv2::Channel *, String> > operators;
-    while(NULL != (pRow = mysql_fetch_row(pRes))) {
-        FTSSrv2::Channel *pChan = this->findChannel(pRow[1]);
-
-        if(!pChan)
-            continue;
-
-        operators.push_back(std::make_pair(pChan, pRow[0]));
-    }
-
-    DataBase::getUniqueDB()->free(pRes);
-
-    // Now we execute that action (only now as the DB result needs to be freed).
-    for(std::list<std::pair<FTSSrv2::Channel *, String> >::iterator i = operators.begin() ; i != operators.end() ; ++i) {
-        i->first->op(i->second, true);
-    }
-
-    return ERR_OK;
-}
-
-int FTSSrv2::ChannelManager::saveChannels( void )
-{
-    Lock l(m_mutex);
-    for( std::list<FTSSrv2::Channel *>::const_iterator i = m_lpChannels.begin() ;
-         i != m_lpChannels.end() ; i++ ) {
-        (*i)->save( );
-    }
-
-    return ERR_OK;
-}
-
-FTSSrv2::Channel *FTSSrv2::ChannelManager::createChannel(const String & in_sName, const Client *in_pCreater, bool in_bPublic)
-{
-    // Every user can only create a limited amount of channels!
-    if(this->countUserChannels(in_pCreater->getNick()) >= DSRV_MAX_CHANS_PER_USER) {
-        return NULL;
-    }
-
-    Lock l(m_mutex);
-    FTSSrv2::Channel *pChannel = new FTSSrv2::Channel(-1, in_bPublic, in_sName,
-                                      DSRV_DEFAULT_MOTTO,
-                                      in_pCreater->getNick());
-
-    m_lpChannels.push_back(pChannel);
-    pChannel->save(); // Update the database right now!
-
-    return pChannel;
-}
-
-int FTSSrv2::ChannelManager::removeChannel(FTSSrv2::Channel *out_pChannel, const String &in_sWhoWantsIt)
-{
-    std::unique_ptr<Lock> l(new Lock(m_mutex));
-    for(std::list<FTSSrv2::Channel *>::iterator i = m_lpChannels.begin() ; i != m_lpChannels.end() ; ++i) {
-        if(*i == out_pChannel) {
-            // Found! remove it from DB and manager, if we have the rights!
-            if(ERR_OK == out_pChannel->destroyDB(in_sWhoWantsIt)) {
-                m_lpChannels.erase(i);
-                // The order here is important, as deleting the channel might kick players, that will be locking the mutex.
-                l.reset();
-                SAFE_DELETE(out_pChannel);
-                return ERR_OK;
-            } else {
-                // Found, but no right to remove it.
-                return -2;
-            }
-        }
-    }
-
-    // Not found.
-    return -1;
-}
-
-std::list<FTSSrv2::Channel *> FTSSrv2::ChannelManager::getPublicChannels()
-{
-    std::list<FTSSrv2::Channel *>lpPubChannels;
-
-    Lock l(m_mutex);
-    for( std::list<FTSSrv2::Channel *>::const_iterator i = m_lpChannels.begin() ;
-         i != m_lpChannels.end() ; i++ ) {
-        if( (*i)->isPublic( ) ) {
-            lpPubChannels.push_back(*i);
-        }
-    }
-
-    return lpPubChannels;
-}
-
-int FTSSrv2::ChannelManager::joinChannel( FTSSrv2::Channel *out_pChannel, Client *out_pClient )
-{
-    if(out_pChannel == NULL)
-        return -1;
-
-    FTSSrv2::Channel *pOldChan = out_pClient->getMyChannel();
-
-    Lock l(m_mutex);
-    // Leave the old channel.
-    if(pOldChan) {
-        pOldChan->quit(out_pClient);
-    }
-
-    // Join the new channel.
-    out_pChannel->join(out_pClient);
-
-    return ERR_OK;
-}
-
-FTSSrv2::Channel *FTSSrv2::ChannelManager::findChannel(const String & in_sName)
-{
-    Lock l(m_mutex);
-    for(std::list<FTSSrv2::Channel *>::const_iterator i = m_lpChannels.begin() ; i != m_lpChannels.end() ; ++i) {
-        // "Pompei2's ChanNel" is the same as "pOmpei2's chAnnEl"
-        if((*i)->getName().ieq(in_sName)) {
-            return *i;
-        }
-    }
-
-    return NULL;
-}
-
-uint32_t FTSSrv2::ChannelManager::countUserChannels(const String &in_sUserName)
-{
-    uint32_t nChans = 0;
-
-    Lock l(m_mutex);
-    for(std::list<FTSSrv2::Channel *>::const_iterator i = m_lpChannels.begin() ; i != m_lpChannels.end() ; ++i) {
-        if((*i)->getAdmin().ieq(in_sUserName)) {
-            nChans++;
-        }
-    }
-
-    return nChans;
-}
-
-std::list<String> FTSSrv2::ChannelManager::getUserChannels(const String &in_sUserName)
-{
-    std::list<String>sChans;
-
-    Lock l(m_mutex);
-    for(std::list<FTSSrv2::Channel *>::const_iterator i = m_lpChannels.begin() ; i != m_lpChannels.end() ; ++i) {
-        if((*i)->getAdmin().ieq(in_sUserName)) {
-            sChans.push_back((*i)->getName());
-        }
-    }
-
-    return sChans;
-}
-
-FTSSrv2::Channel *FTSSrv2::ChannelManager::getDefaultChannel(void)
-{
-    return this->findChannel(DSRV_DEFAULT_CHANNEL_NAME);
+    return nullptr;
 }
