@@ -1,8 +1,13 @@
 #include "server_log.h"
 #include "db.h"
 #include "constants.h"
+#if defined(_MSC_VER)
+#include <mysql.h>
+#include <errmsg.h>
+#else
+#include <mysql/mysql.h>
 #include <mysql/errmsg.h>
-
+#endif
 using namespace FTS;
 using namespace FTSSrv2;
 
@@ -13,7 +18,7 @@ FTS::String FTSSrv2::DataBase::m_psTblFeedbackFields[DSRV_TBL_FEEDBACK_COUNT];
 
 FTSSrv2::DataBase::DataBase()
 {
-    m_pSQL = NULL;
+    m_pSQL = nullptr;
 
     // Fill the arrays with the table names.
     m_psTblUsrFields[DSRV_TBL_USR_ID]="id";
@@ -60,13 +65,13 @@ FTSSrv2::DataBase::~DataBase()
 {
     if(m_pSQL) {
         mysql_close(m_pSQL);
-        m_pSQL = NULL;
+        m_pSQL = nullptr;
     }
 }
 
 int FTSSrv2::DataBase::init()
 {
-    if(NULL == (m_pSQL = mysql_init(NULL))) {
+    if(nullptr == (m_pSQL = mysql_init(nullptr))) {
         FTSMSG("[ERROR] MySQL init: "+this->getError(), MsgType::Error);
         return -1;
     }
@@ -75,12 +80,12 @@ int FTSSrv2::DataBase::init()
     mysql_options(m_pSQL, MYSQL_OPT_RECONNECT, &bTrue);
     mysql_options(m_pSQL, MYSQL_OPT_COMPRESS, &bTrue);
 
-    if(NULL == mysql_real_connect(m_pSQL, DSRV_MYSQL_HOST, DSRV_MYSQL_USER,
+    if(nullptr == mysql_real_connect(m_pSQL, DSRV_MYSQL_HOST, DSRV_MYSQL_USER,
                                   DSRV_MYSQL_PASS, DSRV_MYSQL_DB, 0, NULL,
                                   CLIENT_MULTI_STATEMENTS)) {
         FTSMSG("[ERROR] MySQL connect: "+this->getError(), MsgType::Error);
         mysql_close(m_pSQL);
-        m_pSQL = NULL;
+        m_pSQL = nullptr;
         return -2;
     }
 
@@ -104,7 +109,7 @@ int FTSSrv2::DataBase::free(MYSQL_RES *&out_pRes)
     if(out_pRes) {
         mysql_free_result(out_pRes);
     }
-    out_pRes = NULL;
+    out_pRes = nullptr;
 
     // And then every other if there are some more from multiple queries:
     while(mysql_next_result(m_pSQL) == 0) {
@@ -126,9 +131,22 @@ bool FTSSrv2::DataBase::query(MYSQL_RES *&out_pRes, const FTS::String & in_sStr)
     return this->query(out_pRes, in_sStr.c_str(), NULL);
 }
 
-FTS::String FTSSrv2::DataBase::escape(const FTS::String & in_sStr)
+/// Returns the MySQL escaped version of this string.
+/// this creates a string that is a copy of this one, but escaped for mysql commands.
+///
+/// \param in_sStr The string to escape.
+///
+/// \return The MySQL escaped copy of this string.
+FTS::String FTSSrv2::DataBase::escape( const FTS::String & in_sStr )
 {
-    return in_sStr.mysqlEscaped(m_pSQL);
+    char *buf = new char[in_sStr.len() * 2 + 1];
+    int iLen = mysql_real_escape_string( m_pSQL, buf, in_sStr.c_str(), in_sStr.len() );
+
+    String sRet( buf, 0, iLen );
+    delete[] buf;
+
+    return sRet;
+
 }
 
 int FTSSrv2::DataBase::storedFunctionInt(const FTS::String & in_sFunc, const FTS::String & in_sArgs)
@@ -151,11 +169,11 @@ bool FTSSrv2::DataBase::storedProcedure(MYSQL_RES *&out_pRes, const FTS::String 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! //
 bool FTSSrv2::DataBase::query(MYSQL_RES *&out_pRes, const char *in_pszQuery, ...)
 {
-    out_pRes = NULL;
+    out_pRes = nullptr;
 
     va_list args;
     const char *pszArg = in_pszQuery;
-    char *pszQuery = NULL;
+    char *pszQuery = nullptr;
 
     int n = 0, i = 0;
 
@@ -176,7 +194,7 @@ bool FTSSrv2::DataBase::query(MYSQL_RES *&out_pRes, const char *in_pszQuery, ...
 
     // If the result has to be freed, this mutex only gets unlocked when calling free.
     m_mutex.lock();
-
+    FTSMSGDBG( "SQL Query : " + String(pszQuery), 4 );
     // Execute the query.
     if(0 != mysql_query(m_pSQL, pszQuery)) {
         FTSMSG("[ERROR] MySQL query\nQuery string: "+String(pszQuery)+"\nError: "+this->getError(), MsgType::Error);
@@ -193,7 +211,7 @@ bool FTSSrv2::DataBase::query(MYSQL_RES *&out_pRes, const char *in_pszQuery, ...
     ::free(pszQuery);
 
     out_pRes = mysql_store_result(m_pSQL);
-    if(out_pRes == NULL && mysql_field_count(m_pSQL) != 0) {
+    if(out_pRes == nullptr && mysql_field_count(m_pSQL) != 0) {
         // Errors occured during the query. (connection ? too large result ?)
         FTSMSG("[ERROR] MySQL field count\nQuery string: "+String(pszQuery)+"\nError: "+this->getError(), MsgType::Error);
         return false;
@@ -204,16 +222,16 @@ bool FTSSrv2::DataBase::query(MYSQL_RES *&out_pRes, const char *in_pszQuery, ...
 
 int FTSSrv2::DataBase::storedFunctionInt(const char *in_pszFunc, const char *in_pszArgs)
 {
-    MYSQL_RES *pRes = NULL;
-    MYSQL_ROW row = NULL;
+    MYSQL_RES *pRes = nullptr;
+    MYSQL_ROW row = nullptr;
 
     // There was an error hum hum ...
-    if(!this->query(pRes, "SELECT `"DSRV_MYSQL_DB"`.`", in_pszFunc, "` ( ", in_pszArgs, " )", NULL)) {
+    if(!this->query(pRes, "SELECT `" DSRV_MYSQL_DB "`.`", in_pszFunc, "` ( ", in_pszArgs, " )", NULL)) {
         return -1;
     }
 
     // How can it happen that a stored function returns nothing ??
-    if(NULL == (row = mysql_fetch_row(pRes))) {
+    if(nullptr == (row = mysql_fetch_row(pRes))) {
         this->free(pRes);
         FTSMSG("[ERROR] MySQL fetch stored function "+String(in_pszFunc)+"\nArguments: "+String(in_pszArgs)+"\nError: "+this->getError(), MsgType::Error);
         return -1;
@@ -226,7 +244,7 @@ int FTSSrv2::DataBase::storedFunctionInt(const char *in_pszFunc, const char *in_
 
 bool FTSSrv2::DataBase::storedProcedure(MYSQL_RES *&out_pRes, const char *in_pszProc, const char *in_pszArgs)
 {
-    return this->query(out_pRes, "CALL `"DSRV_MYSQL_DB"`.`", in_pszProc, "` ( ", in_pszArgs, " )", NULL);
+    return this->query(out_pRes, "CALL `" DSRV_MYSQL_DB "`.`", in_pszProc, "` ( ", in_pszArgs, " )", NULL);
 }
 
 FTS::String FTSSrv2::DataBase::getError()
@@ -234,7 +252,7 @@ FTS::String FTSSrv2::DataBase::getError()
     return FTS::String(mysql_error(m_pSQL));
 }
 
-static FTSSrv2::DataBase *g_pTheDatabase = NULL;
+static FTSSrv2::DataBase *g_pTheDatabase = nullptr;
 
 int FTSSrv2::DataBase::initUniqueDB()
 {
