@@ -1,3 +1,5 @@
+#include <type_traits>
+
 #include "internaltester.h"
 #include "logging/logger.h"
 #include "dLib/dArchive/dArchive.h"
@@ -5,6 +7,7 @@
 
 using namespace FTSArc;
 using namespace FTS;
+using namespace std;
 
 InternalTester::InternalTester()
 {
@@ -43,68 +46,68 @@ int InternalTester::execute()
 
     FTSMSG("Reading arbitrary 8-bit chunk:                 ");
     scdc.setCursorPos(5);
-    if(scdc.readi8() != 5) {
+    if(scdc.read<std::int8_t>() != 5) {
         FTSMSG("FAIL\n");
     }
     scdc.setCursorPos(0);
     FTSMSG("OK\n");
 
     FTSMSG("Reading signed 8-bit chunk:                    ");
-    if(scdc.readi8() != 0) {
+    if(scdc.read<std::int8_t>() != 0) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading unsigned 8-bit chunk:                  ");
-    if(scdc.readui8() != 1) {
+    if(scdc.read<std::uint8_t>() != 1) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading signed 16-bit chunk:                   ");
-    if(scdc.readi16() != 770) {
+    if(scdc.read<std::int16_t>() != 770) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading unsigned 16-bit chunk:                 ");
-    if(scdc.readui16() != 1284) {
+    if(scdc.read<std::uint16_t>() != 1284) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading signed 32-bit chunk:                   ");
-    if(scdc.readi32() != 151521030) {
+    if(scdc.read<std::int32_t>() != 151521030) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading unsigned 32-bit chunk:                 ");
-    if(scdc.readui32() != 218893066) {
+    if(scdc.read<std::uint32_t>() != 218893066) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading signed 64-bit chunk:                   ");
-    if(scdc.readi64() != 1518859942647303950) {
+    if(scdc.read<std::int64_t>() != 1518859942647303950) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading unsigned 64-bit chunk:                 ");
-    if(scdc.readui64() != 2097581325351917334) {
+    if(scdc.read<std::uint64_t>() != 2097581325351917334) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading 32-bit floating point chunk:           ");
-    if(scdc.readf() - 5.42512919e-19 >= 0.1e-19) {
+    if(scdc.read<std::float_t>() - 5.42512919e-19 >= 0.1e-19) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
 
     FTSMSG("Reading 64-bit floating point chunk:           ");
-    if(scdc.readd() - 2.008636e-110 >= 0.1e-110) {
+    if(scdc.read<std::double_t>() - 2.008636e-110 >= 0.1e-110) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -120,7 +123,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back signed 8-bit chunk:          ");
     sdc.insert(static_cast<int8_t>(-2));
     sdc.moveCursor(-1);
-    if(sdc.readi8() != -2) {
+    if(sdc.read<std::int8_t>() != -2) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -128,7 +131,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back unsigned 8-bit chunk:        ");
     sdc.insert(static_cast<uint8_t>(252u));
     sdc.moveCursor(-1);
-    if(sdc.readui8() != 252u) {
+    if(sdc.read<std::uint8_t>() != 252u) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -136,7 +139,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back signed 16-bit chunk:         ");
     sdc.insert(static_cast<int16_t>(-3));
     sdc.moveCursor(-2);
-    if(sdc.readi16() != -3) {
+    if(sdc.read<std::int16_t>() != -3) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -144,7 +147,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back unsigned 16-bit chunk:       ");
     sdc.insert(static_cast<uint16_t>(64574u));
     sdc.moveCursor(-2);
-    if(sdc.readui16() != 64574u) {
+    if(sdc.read<std::uint16_t>() != 64574u) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -152,7 +155,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back signed 32-bit chunk:         ");
     sdc.insert(static_cast<int32_t>(-4));
     sdc.moveCursor(-4);
-    if(sdc.readi32() != -4) {
+    if(sdc.read<std::int32_t>() != -4) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -160,7 +163,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back unsigned 32-bit chunk:       ");
     sdc.insert(static_cast<uint32_t>(4194957299u));
     sdc.moveCursor(-4);
-    if(sdc.readui32() != 4194957299u) {
+    if(sdc.read<std::uint32_t>() != 4194957299u) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -168,7 +171,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back signed 64-bit chunk:         ");
     sdc.insert(static_cast<int64_t>(-5));
     sdc.moveCursor(-8);
-    if(sdc.readi64() != -5) {
+    if(sdc.read<std::int64_t>() != -5) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -176,7 +179,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back unsigned 64-bit chunk:       ");
     sdc.insert(static_cast<uint64_t>(17446746073709651614u));
     sdc.moveCursor(-8);
-    if(sdc.readui64() != 17446746073709651614u) {
+    if(sdc.read<std::uint64_t>() != 17446746073709651614u) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -184,7 +187,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back 32-bit floating point chunk: ");
     sdc.insert(static_cast<float>(-1.1f));
     sdc.moveCursor(-4);
-    if(sdc.readf() != -1.1f) {
+    if(sdc.read<std::float_t>() != -1.1f) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -192,7 +195,7 @@ int InternalTester::execute()
     FTSMSG("Write & Read back 64-bit floating point chunk: ");
     sdc.insert(static_cast<double>(-1.1));
     sdc.moveCursor(-8);
-    if(sdc.readd() != -1.1) {
+    if(sdc.read<std::double_t>() != -1.1) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -204,10 +207,10 @@ int InternalTester::execute()
     sdc.setCursorPos(2);
     sdc.insert(static_cast<uint32_t>(4294967294u));
     sdc.moveCursor(-4);
-    if(sdc.readui32() != 4294967294u) {
+    if(sdc.read<std::uint32_t>() != 4294967294u) {
         FTSMSG("FAIL\n");
     }
-    if(sdc.readi16() != -3) {
+    if(sdc.read<std::int16_t>() != -3) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -216,10 +219,10 @@ int InternalTester::execute()
     sdc.setCursorPos(0);
     sdc.insert(static_cast<uint32_t>(4294967293u));
     sdc.moveCursor(-4);
-    if(sdc.readui32() != 4294967293u) {
+    if(sdc.read<std::uint32_t>() != 4294967293u) {
         FTSMSG("FAIL\n");
     }
-    if(sdc.readi8() != -2) {
+    if(sdc.read<std::int8_t>() != -2) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -227,7 +230,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back signed 8-bit chunk:          ");
     sdc.overwrite(static_cast<int8_t>(-22));
     sdc.moveCursor(-1);
-    if(sdc.readi8() != -22) {
+    if(sdc.read<std::int8_t>() != -22) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -235,7 +238,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back unsigned 8-bit chunk:        ");
     sdc.overwrite(static_cast<uint8_t>(152u));
     sdc.moveCursor(-1);
-    if(sdc.readui8() != 152u) {
+    if(sdc.read<std::uint8_t>() != 152u) {
         FTSMSG("FAIL\n");
     }
     sdc.moveCursor(+4);
@@ -244,7 +247,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back signed 16-bit chunk:         ");
     sdc.overwrite(static_cast<int16_t>(-3));
     sdc.moveCursor(-2);
-    if(sdc.readi16() != -3) {
+    if(sdc.read<std::int16_t>() != -3) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -252,7 +255,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back unsigned 16-bit chunk:       ");
     sdc.overwrite(static_cast<uint16_t>(64574u));
     sdc.moveCursor(-2);
-    if(sdc.readui16() != 64574u) {
+    if(sdc.read<std::uint16_t>() != 64574u) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -260,7 +263,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back signed 32-bit chunk:         ");
     sdc.overwrite(static_cast<int32_t>(-4));
     sdc.moveCursor(-4);
-    if(sdc.readi32() != -4) {
+    if(sdc.read<std::int32_t>() != -4) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -268,7 +271,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back unsigned 32-bit chunk:       ");
     sdc.overwrite(static_cast<uint32_t>(4194957299u));
     sdc.moveCursor(-4);
-    if(sdc.readui32() != 4194957299u) {
+    if(sdc.read<std::uint32_t>() != 4194957299u) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -276,7 +279,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back signed 64-bit chunk:         ");
     sdc.overwrite(static_cast<int64_t>(-5));
     sdc.moveCursor(-8);
-    if(sdc.readi64() != -5) {
+    if(sdc.read<std::int64_t>() != -5) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -284,7 +287,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back unsigned 64-bit chunk:       ");
     sdc.overwrite(static_cast<uint64_t>(17446746073709651614u));
     sdc.moveCursor(-8);
-    if(sdc.readui64() != 17446746073709651614u) {
+    if(sdc.read<std::uint64_t>() != 17446746073709651614u) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -292,7 +295,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back 32-bit floating point chunk: ");
     sdc.overwrite(static_cast<float>(-1.1f));
     sdc.moveCursor(-4);
-    if(sdc.readf() != -1.1f) {
+    if(sdc.read<std::float_t>() != -1.1f) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -300,7 +303,7 @@ int InternalTester::execute()
     FTSMSG("Overwrite & Read back 64-bit floating point chunk: ");
     sdc.overwrite(static_cast<double>(-1.1));
     sdc.moveCursor(-8);
-    if(sdc.readd() != -1.1) {
+    if(sdc.read<std::double_t>() != -1.1) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
@@ -312,7 +315,7 @@ int InternalTester::execute()
     sdc.moveCursor(-4);
     sdc.overwrite(static_cast<double>(-1.23456789));
     sdc.moveCursor(-8);
-    if(sdc.readd() != -1.23456789) {
+    if(sdc.read<std::double_t>() != -1.23456789) {
         FTSMSG("FAIL\n");
     }
     FTSMSG("OK\n");
